+++ packages/integration/src/addFileScope.test.ts
@@ -1,4 +1,4 @@
-import { outdent } from 'outdent';
+import dedent from 'dedent';
 import { sep, posix, win32 } from 'path';
 
 import { addFileScope, normalizePath } from './addFileScope';
@@ -13,7 +13,7 @@ expect.addSnapshotSerializer({
 
 describe('ESM', () => {
   test('should add missing fileScope', () => {
-    const source = outdent`
+    const source = dedent`
     import {style} from '@vanilla-extract/css';
 
     export const myStyle = style({});
@@ -37,7 +37,7 @@ describe('ESM', () => {
   });
 
   test('should add global adapter setup when "globalAdapterIdentifier" is provided', () => {
-    const source = outdent`
+    const source = dedent`
     import {style} from '@vanilla-extract/css';
 
     export const myStyle = style({});
@@ -65,7 +65,7 @@ describe('ESM', () => {
   });
 
   test('should update existing fileScope', () => {
-    const source = outdent`
+    const source = dedent`
     import { setFileScope, endFileScope } from "@vanilla-extract/css/fileScope";
     setFileScope("some-weird-file", "some-weird-package");
     import {style} from '@vanilla-extract/css';
@@ -92,7 +92,7 @@ describe('ESM', () => {
   });
 
   test('should update existing fileScope with newlines', () => {
-    const source = outdent`
+    const source = dedent`
       import { setFileScope, endFileScope } from "@vanilla-extract/css/fileScope";
       setFileScope(
         "some-weird-file",
@@ -122,7 +122,7 @@ describe('ESM', () => {
   });
 
   test('should handle namespaced filescope calls', () => {
-    const source = outdent`
+    const source = dedent`
     import * as vanillaFileScope from "@vanilla-extract/css/fileScope";
     vanillaFileScope.setFileScope("some-weird-file");
     import {style} from '@vanilla-extract/css';
@@ -151,7 +151,7 @@ describe('ESM', () => {
 
 describe('CJS', () => {
   test('should add missing fileScope', () => {
-    const source = outdent`
+    const source = dedent`
     const _css = require('@vanilla-extract/css');
 
     var myStyle = _css.style({});
@@ -177,7 +177,7 @@ describe('CJS', () => {
   });
 
   test('should add global adapter setup when "globalAdapterIdentifier" is provided', () => {
-    const source = outdent`
+    const source = dedent`
     const _css = require('@vanilla-extract/css');
 
     var myStyle = _css.style({});
@@ -207,7 +207,7 @@ describe('CJS', () => {
   });
 
   test('should update existing fileScope', () => {
-    const source = outdent`
+    const source = dedent`
     const __vanilla_filescope__ = require("@vanilla-extract/css/fileScope");
     __vanilla_filescope__.setFileScope("some-weird-file", "some-weird-package");
     const _css = require('@vanilla-extract/css');
@@ -236,7 +236,7 @@ describe('CJS', () => {
   });
 
   test('should update existing fileScope with newlines', () => {
-    const source = outdent`
+    const source = dedent`
       const __vanilla_filescope__ = require("@vanilla-extract/css/fileScope");
       __vanilla_filescope__.setFileScope(
         "some-weird-file",
@@ -280,7 +280,7 @@ test('platform-specific relative path', () => {
     },
   }[sep];
 
-  const source = outdent`
+  const source = dedent`
     import { style } from '@vanilla-extract/css';
 
     export const myStyle = style({});