+++ fixtures/low-level/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fixtures/low-level
 
+## 0.0.26
+
+### Patch Changes
+
+- Updated dependencies [[`df9fe3ee3fc0057bc14a2333a405f8229f80c214`](https://github.com/vanilla-extract-css/vanilla-extract/commit/df9fe3ee3fc0057bc14a2333a405f8229f80c214)]:
+  - @vanilla-extract/css@1.15.0
+
 ## 0.0.25
 
 ### Patch Changes